requirements.txt
Flask

gunicorn

Flask-SQLAlchemy

postgres

python-telegram-bot

requests
